Fixes of reports

+++ net.ostis.confman.ui/src/net/ostis/confman/ui/reports/ReportReviewingPart.java
@@ -180,8 +180,10 @@ public class ReportReviewingPart {
         if (currentReport != null) {
             currentReport.setReportAccepted(false);
             currentReport.setReportCanceled(true);
+            if(currentReport.getSection() != null && currentReport.getSection().getConference() != null) {
+                conferenceService.deleteReport(currentReport);
+            }
         }
-        conferenceService.deleteReport(currentReport);
     }
 
     private void save() {
